Clean up and reorganize data collection
