+++ pom.xml
@@ -59,7 +59,7 @@
         <version>2.2.0</version>
         <configuration>
           <!-- TODO: set project ID. -->
-          <deploy.projectId>YOUR_PROJECT_ID_HERE</deploy.projectId>
+          <deploy.projectId>step129-2020</deploy.projectId>
           <deploy.version>1</deploy.version>
         </configuration>
       </plugin>
